feat: Prepare application for production deployment

+++ models/budget.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, String, ForeignKey
+from sqlalchemy import Column, Integer, Float, String, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from db.base import Base
 
@@ -8,6 +8,6 @@ class Budget(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"))
     category = Column(String, nullable=False)
-    monthly_limit = Column(Float, nullable=False)
+    monthly_limit = Column(Numeric(10, 2), nullable=False)
 
     user = relationship("User", back_populates="budgets")
